Haha, this feels fragile - but it works.
The danger here is that if you pass in a container that is
neither an iterator itself (supporting `*(x + index)` and `x[index]`
for both read and write) nor supporting `std::begin()` then
the it'll fail with an inscrutable error.

The *worse* danger is you passing in something which *does*
implement iterator-ish, but in some non-iterator way.